requirements.txt
tensorflow==2.9.1
numpy==1.23.5
matplotlib==3.7.1
scikit-learn==1.2.2
tqdm==4.65.0
Pillow==11.2.1
